// src/lib/api/generated/types.ts
export interface components {
    schemas: {
// ...
        ArtistDTOMinified: {
            id?: string;
            name?: string;
            /** Format: double */
            longitude?: number;
            /** Format: double */
            latitude?: number;
            city?: string;
            /** @enum {string} */
            region?: "KYIV_CITY" | "SEVASTOPOL" | "CRIMEA" | "LVIV" | "VOLYN" | "RIVNE" | "TERNOPIL" | "IVANO_FRANKIVSK" | "ZAKARPATTIA" | "CHERNIVTSI" | "KHMELNYTSKYI" | "ZHYTOMYR" | "KYIV" | "CHERNIHIV" | "SUMY" | "VINNYTSIA" | "CHERKASY" | "POLTAVA" | "ODESA" | "MYKOLAIV" | "KHERSON" | "KIROVOHRAD" | "DNIPROPETROVSK" | "KHARKIV" | "ZAPORIZHZHIA" | "DONETSK" | "LUHANSK";
            slug?: string;
            imageUrls?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}